requirements.txt
notion-client==2.2.1
anthropic==0.40.0
python-dotenv==1.0.0
